requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.10.0
astroid==3.3.11
cachetools==6.2.1
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
dill==0.4.0
distro==1.9.0
Django==5.2.8
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
google-auth==2.43.0
google-genai==1.54.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
isort==6.1.0
mccabe==0.7.0
mysqlclient==2.2.7
platformdirs==4.5.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
pylint==3.3.9
pylint-django==2.6.1
pylint-plugin-utils==0.9.0
PyMySQL==1.1.2
pyparsing==3.2.5
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
sqlparse==0.5.3
tenacity==9.1.2
tomlkit==0.13.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
websockets==15.0.1
